upd: hw2 done
