+++ controllers/contacts-operations/refreshContacts.js
@@ -1,0 +1,12 @@
+const fs = require("fs/promises");
+const { contactsPath } = require("../../model");
+
+const refreshContacts = async (contactsToFile) => {
+  const newContacts = await fs.writeFile(
+    contactsPath,
+    JSON.stringify(contactsToFile)
+  );
+  return newContacts;
+};
+
+module.exports = refreshContacts;
