+++ controllers/contacts/index.js
@@ -1,0 +1,5 @@
+const contacts = require("./contacts");
+
+module.exports = {
+  contacts,
+};
