+++ middlewares/index.js
@@ -1,0 +1,7 @@
+const controllerWrapper = require("./controllerWrapper");
+const validation = require("./validation");
+
+module.exports = {
+  controllerWrapper,
+  validation,
+};
