+++ routes/api/contacts.js
@@ -1,103 +1,23 @@
 const express = require("express");
-const { NotFound, BadRequest } = require("http-errors");
 
 const router = express.Router();
-const Joi = require("joi");
 
-const joiSchema = Joi.object({
-  name: Joi.string().min(3).max(30).required(),
-  phone: Joi.string().required(),
-  email: Joi.string().email({ minDomainSegments: 2 }).required(),
-});
+const { contacts: ctrl } = require("../../controllers/contacts");
+const { controllerWrapper, validation } = require("../../middlewares");
+const contactSchema = require("../../schemas");
 
-const contactsOperations = require("../../model");
+router.get("/", controllerWrapper(ctrl.listContacts));
 
-router.get("/", async (req, res, next) => {
-  try {
-    const contacts = await contactsOperations.listContacts();
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        contacts,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", controllerWrapper(ctrl.getContactById));
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    console.log("id is:", contactId);
-    const contact = await contactsOperations.getContactById(contactId);
-    if (!contact) {
-      throw new NotFound(`Contact with id: ${contactId} was not found`);
-      //throw new createError(404, `Contact with id: ${contactId} was not found`);
-      // res.status(404).json({
-      //   status: "error",
-      //   code: 404,
-      //   message: `Contact with id: ${contactId} was not found`,
-      // });
-      // return;
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        contact,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", validation(contactSchema), controllerWrapper(ctrl.addContact));
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = joiSchema.validate(req.body);
-    if (error) {
-      throw new BadRequest(error.message);
-    }
-    const result = await contactsOperations.addContact(req.body);
-    res.status(201).json({
-      status: "success",
-      code: 201,
-      data: {
-        result,
-      },
-    });
-  } catch (error) {}
-  next(error);
-});
+router.delete("/:contactId", controllerWrapper(ctrl.removeContact));
 
-router.delete("/:contactId", async (req, res, next) => {
-  res.json({ message: "template message" });
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = joiSchema.validate(req.body);
-    if (error) {
-      throw new BadRequest(error.message);
-    }
-    const { contactId } = req.params;
-
-    const result = await contactsOperations.updateContact(contactId, req.body);
-    if (!result) {
-      throw new NotFound(`Contact with id: ${contactId} was not found`);
-    }
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        result,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
+router.put(
+  "/:contactId",
+  validation(contactSchema),
+  controllerWrapper(ctrl.updateContact)
+);
 
 module.exports = router;
