+++ schemas/index.js
@@ -1,0 +1,3 @@
+const contactSchema = require("./contact");
+
+module.exports = contactSchema;
